Changes related to 2015 tests.
1-
In the onLoad method of EMEHandler the session variable is used as the session object with MediaKeys or as a session id in the case of 0.1b.
When binding the method we only pass it as id so the session.update call will fail unless we pass the proper object. The patch fixes that.

2-
If you have available both v0.1b and the OO APIs at the same time, setupBaseEmeTest will detect the second and as we fixed previously, we will be sending the session object instead of the necessary session id.
The test could succeed with a V0.1b only browser, but not with one supporting both, specially after the fix for session object.

3-
encryptedMediaPortability.js is now unique to each test suite.
Fixed an issue regarding pre-fixing in 2015 if browser supports v0.1b and the OO APIs.