Handle QuotaExceededErr exception from appendBuffer
New version of Chrome (starting with Chrome 46) will start throwing
QuotaExceededErr exception when SourceBuffer is full, instead of
discarding old data. See this Chromium CL for more details:
https://codereview.chromium.org/1008463002/
So now we need to adjust the logic for detecting buffer fullness:
we know that the buffer is full when either we get a QuotaExceededErr
exception from appendBuffer (Chrome 46 and hopefully other newer
browsers) or when we detect that some data has been discarded
(Chrome older than 46 and other current/older browsers).