Support DOMHighResTimeStamp values in Event.timestamp
This fixes test "51. EventTimestamp" in the 2016 MSE test suite on browsers
implementing high resolution timestamps, as specified in[1]. Those timestamps
don't count from the epoch, but from the time origin[2].

This new behaviour is different from the specified in DOM4[3], where
Event.timestamp is of type DOMTimeStamp and is clearly stated that the epoch
is used as reference[4].

Most modern browsers (Firefox, Chrome, Edge and WebKit) have already migrated
to high res timestamps.

[1] https://dom.spec.whatwg.org/#interface-event
[2] https://w3c.github.io/hr-time/#time-origin
[3] https://www.w3.org/TR/domcore/#event
[4] https://www.w3.org/TR/domcore/#dom-event-timestamp