Change the result report function to new protocal.
This function currently dumps results to qual-e in simple string and no one actually uses it.
But it will be used in future and therefore we need to make a rule about how to report the results.
The test results should be wrapped into an Object in
format of
{
  "pass" : {
    testSuite : {
      testType : {
        category : [
	  test1, test2
	]
      }
    }
  }
  "fail" : {...}
}

For example,
{
  "pass" : {
    "2018" : {
      "conformance-test" : {
	"Media Element Core": [
	  "InitialMediaDuration", "InitialMediaVideoWidth"
	]
      }
    }
  }
}

This would be helpful in automatically using the test results during certification.

Bug: 78533407
Change-Id: I902134a71598ed57b71ed3338090c07a73961eb0